% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careflowMining.R
\name{careflowMining}
\alias{careflowMining}
\title{A class for careflow mining}
\usage{
careflowMining(parameters.list = list())
}
\description{
This is an implementation of the Care Flow Mining algorithm for Process Mining issues.
               This class provides a minimal set of methods to handle with the FOMM model:
             In order to better undestand the use of such methods, please visit: www.pminer.info

             The consturctor admit the following parameters: is a boolean which indicates if the autoloops have to be admitted, while 'threshold' is the minimum value that a probability should have to do not be set to zero, in the transition matrix.
}
