% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_001.R
\name{resource_001}
\alias{resource_001}
\title{A class born to cope with resources}
\usage{
resource_001(parameters.list = list(), verbose.mode = TRUE)
}
\arguments{
\item{parameters.list}{list. a list containing the parameters. The possible ones are: 'considerAutoLoop' and 'threshold'. 'considerAutoLoop' is a boolean which indicates if the autoloops have to be admitted, while 'threshold' is the minimum value that a probability should have to do not be set to zero, in the transition matrix.}

\item{verbose.mode}{boolean. Verbose mode TRUE / FALSE}
}
\description{
A class to manage resourcese
}
